switch to 16 bit transfers

+++ inc/pool.h
@@ -11,7 +11,7 @@
 #define POOL_BUFF_SIZE  200
 #define POOL_NUM_BUFFERS 8
 
-typedef uint8_t pool_item_t;
+typedef uint16_t pool_item_t;
 
 pool_item_t * pool_malloc_buff(void);
 void pool_free_buff(pool_item_t * handle);
